jkgls

+++ videoplay.py
@@ -18,36 +18,47 @@
 
 '''	Import relevant libraries '''
 import Rpi.GPIO as GPIO
-import sysimport os
+import sys
+import os
 from subprocess import Popen
 
 '''	Setup GPIO pins on the Raspberry Pi '''
 GPIO.setmode(GPIO.BCM)
-GPIO.setup(17, GPIO.IN, pull_up_down=GPIO.PUD_UP)
+GPIO.setup(14, GPIO.IN, pull_up_down=GPIO.PUD_UP)
+GPIO.setup(15, GPIO.IN, pull_up_down=GPIO.PUD_UP)
 GPIO.setup(18, GPIO.IN, pull_up_down=GPIO.PUD_UP)
+GPIO.setup(23, GPIO.IN, pull_up_down=GPIO.PUD_UP)
 GPIO.setup(24, GPIO.IN, pull_up_down=GPIO.PUD_UP)
 
 '''	Set up movie locations and names '''	
 movie1 = ("/home/pi/Videos/movie1.mp4")
 movie2 = ("/home/pi/Videos/movie2.mp4")
+movie3 = ("/home/pi/Videos/movie3.mp4")
+movie4 = ("/home/pi/Videos/movie4.mp4")
 
 '''	Set up state variables for the videos and GPIO pins '''
 last_state1 = True
 last_state2 = True
+last_state3 = True
+last_state4 = True
 
 input_state1 = True
 input_state2 = True
+input_state3 = True
+input_state4 = True
 
 quit_video = True
 
 ''' Main loop '''
 while True:
 	'''Read states of inputs'''
-	input_state1 = GPIO.input(17)
-	input_state2 = GPIO.input(18)
-	quite_video = GPIO.input(24)
+	input_state1 = GPIO.input(14)
+	input_state2 = GPIO.input(15)
+	input_state3 = GPIO.input(18)
+	input_state4 = GPIO.input(23)
+	quit_video = GPIO.input(24)
 
-	'''If GPIO(17) is shorted to ground'''
+	'''If GPIO(14) is shorted to ground'''
 	if input_state1 != last_state1:
 		if (player and not input_state1):
 			os.system('killall omxplayer.bin')
@@ -57,7 +68,7 @@ while True:
 			omxc = Popen(['omxplayer', '-b', movie1])
 			player = True
 	
-	'''If GPIO(18) is shorted to ground'''
+	'''If GPIO(15) is shorted to ground'''
 	elif input_state2 != last_state2:
 		if (player and not input_state2):
 			os.system('killall omxplayer.bin')
@@ -66,9 +77,29 @@ while True:
 		elif not input_state2:
 			omxc = Popen(['omxplayer', '-b', movie2])
 			player = True
-
-	'''If omxplayer is running and GPIO(17) and GPIO(18) are NOT shorted to ground'''
-	elif (player and input_state1 and input_state2):
+	
+	'''If GPIO(18) is shorted to ground'''
+	elif input_state3 != last_state3:
+		if (player and not input_state3):
+			os.system('killall omxplayer.bin')
+			omxc = Popen(['omxplayer', '-b', movie3])
+			player = True
+		elif not input_state3:
+			omxc = Popen(['omxplayer', '-b', movie3])
+			player = True	
+	
+	'''If GPIO(23) is shorted to ground'''
+	elif input_state4 != last_state4:
+		if (player and not input_state4):
+			os.system('killall omxplayer.bin')
+			omxc = Popen(['omxplayer', '-b', movie4])
+			player = True
+		elif not input_state4:
+			omxc = Popen(['omxplayer', '-b', movie4])
+			player = True
+	
+	'''If omxplayer is running and GPIO pins are NOT shorted to ground'''
+	elif (player and input_state1 and input_state2 and input_state3 and input_state4):
 		os.system('killall omxplayer.bin')
 		player = False
 
@@ -80,4 +111,8 @@ while True:
 	'''Set last_input states'''
 	last_state1 = input_state1
 	last_state2 = input_state2 
+	last_state3 = input_state3
+	last_state4 = input_state4
+	'''' unnecessa
+
